Factor out Border and BorderRadius into their own files.
This will make it more tractable to convert them for RTL.
